FIR IDE: move common logic in add accessor intention to idea-frontend-independent
